add script

+++ my_model/mingpt/utils.py
@@ -38,7 +38,7 @@ def sample(model, x, steps, temperature=1.0, sample=False, top_k=None):
         else:
             _, ix = tf.math.top_k(probs, k=1)
         # append to the sequence and continue
-        ix = tf.cast(ix, tf.float32)
+        ix = tf.cast(ix, tf.int32)
         x = tf.concat((x,ix), axis=1)
 
     return x
